Help out MSVC with determining the right overload for pow().
git-svn-id: https://svn.salilab.org/imp/trunk@12887 e91f8c7b-0e32-4433-a9ac-4502d39a4eab

+++ modules/isd/include/vonMisesSufficient.h
@@ -113,7 +113,7 @@ class vonMisesSufficient : public base::Object
 
   void set_N(unsigned N){
     N_=N;
-    I0N_=pow(I0_,N_);
+    I0N_=pow(I0_, static_cast<int>(N_));
     logterm_ = log(2*IMP::PI*I0N_);
   }
 
@@ -122,7 +122,7 @@ class vonMisesSufficient : public base::Object
         kappa_ = kappa;
         I0_ = internal::bessel_i0(kappa);
         I1_ = internal::bessel_i1(kappa);
-        I0N_=pow(I0_,N_);
+        I0N_=pow(I0_, static_cast<int>(N_));
         logterm_ = log(2*IMP::PI*I0N_);
     }
   }
